+++ Cargo.toml
@@ -1,10 +1,15 @@
-[package]
-name = "scheduler"
+[workspace]
+resolver = "3"
+members = ["scheduler", "tasks"]
+
+[workspace.package]
 version = "0.1.0"
 edition = "2024"
 
-[dependencies]
+[workspace.dependencies]
 ciborium = "0.2.2"
 serde = { version = "1.0.219", features = ["derive"] }
 thiserror = "2.0.12"
 typetag = "0.2.20"
+
+scheduler = { path = "./scheduler" }
