+++ tasks/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "tasks"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+ciborium.workspace = true
+serde.workspace = true
+thiserror.workspace = true
+typetag.workspace = true
+
+scheduler.workspace = true
